+++ services/madoc-ts/src/frontend/site/hooks/use-relative-links.ts
@@ -17,6 +17,7 @@ export function useRelativeLinks() {
         query?: any;
         subRoute?: string;
         admin?: boolean;
+        hash?: string;
       } = {}
     ) => {
       return createLink({